C - More singly linked lists

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -14,7 +14,7 @@
 size_t print_listint(const listint_t *h)
 {
 	const listint_t *ptr;
-	size_t n;
+	unsigned int n;
 
 	ptr = h;
 	n = 0;
